@@ -17,16 +17,12 @@ Gem::Specification.new do |s|
 
   s.executables   = ["railsdb", "rails_db", 'runsql']
 
-  s.add_dependency 'rails', ">= 3.1.0"
+  s.add_dependency 'rails', ">= 5.0.0"
   s.add_dependency 'codemirror-rails'
   s.add_dependency 'terminal-table'
   s.add_dependency 'simple_form'
   s.add_dependency 'ransack'
-  if RUBY_VERSION == '1.9.3'
-    s.add_dependency 'kaminari', '< 1.0.0'
-  else
-    s.add_dependency 'kaminari'
-  end
+  s.add_dependency 'kaminari'
 
   s.add_development_dependency "launchy"
   s.add_development_dependency "sqlite3"
