@@ -1,4 +1,4 @@
-class PopulateData < ActiveRecord::Migration
+class PopulateData < ActiveRecord::Migration[4.2]
   def change
     # name:string age:integer salary:decimal bio:text
     user_1 = User.create(:name => "John Smith", :age => 30, :salary => 123.45, :bio => 'This is text')
