@@ -1,4 +1,4 @@
-class PopulateDb < ActiveRecord::Migration
+class PopulateDb < ActiveRecord::Migration[4.2]
   def change
   end
 end
