@@ -1,4 +1,4 @@
-class AddNonPkTable < ActiveRecord::Migration
+class AddNonPkTable < ActiveRecord::Migration[4.2]
   def change
     create_table :projects_users, id: false do |t|
       t.integer :project_id
