@@ -1,4 +1,4 @@
-class AddBooleanFieldToUser < ActiveRecord::Migration
+class AddBooleanFieldToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :active, :boolean, default: true
   end
