@@ -1,3 +1,3 @@
 module RailsDb
-  VERSION = "1.6.0"
+  VERSION = "2.0.0"
 end
