@@ -64,26 +64,26 @@ class DashboardControllerTest < ActionDispatch::IntegrationTest
     post '/rails/db/sql-xls.xls?sql=select+%2A+from+users+limit+10'
     assert_equal 200, status
 
-    xhr :get, '/rails/db/tables/users/edit?pk_id=' + edit_user.id.to_s
+    get '/rails/db/tables/users/edit?pk_id=' + edit_user.id.to_s, xhr: true
     assert_equal 200, status
 
-    xhr :put, '/rails/db/tables/users/update?pk_id=' + edit_user.id.to_s, {record: { name: 'JOHN' }}
+    put '/rails/db/tables/users/update?pk_id=' + edit_user.id.to_s, {xhr: true, params: { record: { name: 'JOHN' }}}
     assert_equal 200, status
     edit_user.reload
     assert_equal 'JOHN', edit_user.name
 
-    xhr :get, '/rails/db/tables/users/new'
+    get '/rails/db/tables/users/new', xhr: true
     assert_equal 200, status
 
     assert_equal 2, User.count
-    xhr :post, '/rails/db/tables/users/create', {record: { name: 'XXX' }}
+    post '/rails/db/tables/users/create', {xhr: true, params: { record: { name: 'XXX' }} }
     assert_equal 200, status
     assert_equal 3, User.count
 
-    xhr :get, '/rails/db/tables/users/edit?pk_id=' + customer.id.to_s
+    get '/rails/db/tables/users/edit?pk_id=' + customer.id.to_s, xhr: true
     assert_equal 200, status
 
-    xhr :put, '/rails/db/tables/users/update?pk_id=' + customer.id.to_s, {record: { name: 'STI' }}
+    put '/rails/db/tables/users/update?pk_id=' + customer.id.to_s, {xhr: true, params: { record: { name: 'STI' }} }
     assert_equal 200, status
     customer.reload
     assert_equal 'STI', customer.name
