@@ -22,7 +22,15 @@ Rails.application.configure do
 
   # Disable serving static files from the `/public` folder by default since
   # Apache or NGINX already handles this.
-  config.serve_static_files = true
+  begin
+    config.serve_static_files = true
+  rescue NoMethodError
+  end
+
+  begin
+    config.public_file_server.enabled = true
+  rescue NoMethodError
+  end
 
   config.log_level = :error
 
